Reorganize the calculator demo
This patch is a warmup to improving the visuals in the calculator demo.

Related to #4535
