+++ examples/flutter_gallery/lib/demo/calculator/logic.dart
